2-project-modified/python-client/requirements.txt
confluent-kafka>=2.8.0
python-dotenv
Faker>=19.6.2
pytest
pytest-timeout
pytest-asyncio
jsonschema>=4.0.0
requests>=2.25.0
httpx
cachetools
authlib
referencing
fastavro